+++ packages/common/web/src/consts.js
@@ -76,3 +76,8 @@ export const PLUGIN_DECORATION_PROPS = {
     containerClassName: props.containerClassName,
   }),
 };
+
+export const TOOLBAR_OFFSETS = {
+  top: 12,
+  left: 15,
+};